backspringboot/requirements.txt
fastapi
uvicorn
azure-storage-blob>=12.13.0
pydantic
python-dotenv